+++ lib/screens/word_list/word_list_view_entry_screen.dart
@@ -27,7 +27,8 @@ class WordListViewEntryScreen extends StatelessWidget {
       body: Dictionary(
         false,
         initialEntryId: entry.id,
-        includeFallingWords: false,),
+        includeFallingWords: false,
+      ),
     );
   }
 }